\name{mpra-package}
\alias{mpra-package}
\alias{mpra}
\docType{package}
\title{
\packageTitle{mpra}
}
\description{
\packageDescription{mpra}
}
\details{
  FIXME
}
\author{
\packageAuthor{mpra}

Maintainer: \packageMaintainer{mpra}
}
\references{
  FIXME
}
\keyword{ package }
\examples{
  ## FIXME
}
