\name{mpraSetExample}
\alias{mpraSetExample}
\alias{mpraSetAggExample}
\docType{data}
\title{
  Example data for the mpra package.
}
\description{
  Example data for the MPRA package. The package contains data both at
  the barcode level and aggregated over barcodes. The data is XX.
}
\usage{
data("mpraSetExample")
data("mpraSetAggExample")
}
\format{
  An \code{MPRASet}.
}
\details{
  FIXME
}
\source{
  FIMXE
}
\references{
  FIXME
}
\examples{
data(mpraSetAggExample)
}
\keyword{datasets}
